requirements.txt
bcrypt==4.0.1
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cssselect2==0.7.0
dropbox==11.36.0
Flask==2.2.2
fonttools==4.38.0
gunicorn==20.1.0
html5lib==1.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector==2.2.9
Pillow==9.3.0
ply==3.11
pycparser==2.21
pydyf==0.5.0
pyphen==0.13.2
pysqlcipher3==1.1.0
requests==2.28.1
six==1.16.0
stone==3.3.1
tinycss2==1.2.1
urllib3==1.26.13
weasyprint==57.2
webencodings==0.5.1
Werkzeug==2.2.2
zopfli==0.2.2
